And now the other intel compiler...
Now that we've moved the procedure pointer into the closure so that the new
intel compiler, ifx, works, the old intel compiler, ifort, has started to
have issues with that data_ptr that gets passed around.

The issue was observed while running data_c_exe. It passes a data pointer
and in the callback it checks to make sure that it gets back the same data
pointer. It started complaining.

Removing the value attribute seems to have fixed it, and this seems to work
for other compilers as well.