Reverting the change to the value attribute of type(C_PTR) and other cosmetic fixes
After reviewing the Fortran 2008, 2018, and 2023 standards, it looks like value is meant to
be an attribute for type(C_PTR) in the way that we are using it, so it is probably best
to leave it in and accept the fact that the classic Intel fortran compiler does not work in
this case. It's worth noting that the only case where there is an issue is when sending custom
data to the objective function, which is not a common use case, and hopefully those who need
that functionality are able to use a compiler other than the classic Intel one.

Minimal example: https://fortran-lang.discourse.group/t/problem-with-ifort-compiler-and-type-c-ptr/7011

Other various minor changes are made based on feedback from an in-person review session
with @zaikunzhang