Get the version number using git describe
This lets us avoid having a hardcoded file in the repo and having to
update it for a release. This way releases can be done simply by
tagging the appropriate commit and rebuilding.

This will be useful for when we publish packages, since it lets us publish
a package every time we run CI and not have it conflict with or overwrite
the latest package.