Remove allocations for xl, xu, nlconstr0 in C interface
As I was working on the Python bindings, I ran into some issues related to
memory management. These changes should help with that, particularly the change to
the allocation of nlconstr0, since that was the one that appeared to be giving issues
with Python. In general I'd prefer to do as little memory allocation within the C interface
as possible, since it's challenging enough to have memory management in 1 language, let
alone 4 when we start interacting with Python->C++->C->Fortran.

What enables the removal of these allocations is that I've discovered that if a variable in
Fortran is marked as a pointer and set to null(), then the Fortran function `present`
returns false on that variable. This means we can use NULL from within C to indicate that a
variable is not provided. This approach has a small inconvenience in that the user will
have to handle allocations for optional variables.

For example, with f0, the user would have to do the following:

    double f0 = obj_fun(x0);
    problem->f0 = &f0;

Likewise for things like rhobeg, rhoend, the user would have to do something like the following:

    double rhobeg = 1.0;
    options->rhobeg = &rhobeg;
    double rhoend = 1.0e-6;
    options->rhoend = &rhoend;

This is in constrast to the current way of doing things:

    options->rhobeg = 1.0;
    options->rhoend = 1.0e-6;

As annoying as the previous option is, it might be preferable since it means we can stop
duplicating initialization logic within C. That logic is currently implemented in
`prima_init_options` and `prima_check_problem`, but it doesn't replicate the full logic for
rhobeg/rhoend. I think that ideally C should be a 'dumb' pass-through and should let Fortran
handle as much as possible.

We can use this PR to add this functionality to the rest of the variables, but I figured I
would start with removing the memory allocations before going too much further.