Amending from PR feedback
An intermediate variable was introduced so that the interface will
still work if RP=4.

The method of providing f0 was changed to be a little bit more user
friendly, and rhobeg and rhoend adopted this approach.

Additionally, the "unset" value of maxfun and npt was changed to 0
based on discussion in the PR.