Fixing gdb issue with gcc11 on Windows
It complained "A syntax error in expression, near `= -1'."

Apparently gdb was evaluating the expressions in the command file as if
they were in Fortran. Since Fortran 77 does not have "==" this produced
a syntax error. Setting the language to C manually appears to resolve
this.

It's also not entirely clear why we were looking for exitcode to be
equal to -1. Some tests showed that if the exitcode were 2 this would
be a problem, so we look at the more reasonable value of 0 instead.