Add new words
I'm not sure what the deal is with "inear". I see no instance of such a
word by itself. I saw this once before, but can't remember what the
issue was, or the fix. Maybe there' a line that starts with Linear and
the previous newline is \r\n and screws up the spelling processing. I
don't know, just venturing a guess.