Refactor processing of nl constraints back to not needing x0
Discussed with @zaikunzhang and decided that checking for the size
to be correct every time we execute the constraint function is
preferable since there are edge cases where the constraint function
could return something of a different size. Unlikely, but it is
something we want to catch. Also there will be further processing
inside the constraint function which will likely increase the
performance needs to the point that a could extra if statements would
not be very harmful.