Run profiles against MATLAB version of PRIMA
The profiles script has been modified to test the Python version of
PRIMA against the MATLAB version by way of the MATLAB engine for Python
and sockets. Since the MATLAB version has been quite thoroughly vetted,
if the Python version compares favorably to it then by extension this
gives us confidence in the Python verison.