Fix bug where COBYLA could not run w/o NL constraints
Also added tests for it. Also removed a comment that referred to a
previous way of doing thing, and also made the method names lowercase
since users may look at these tests as a guide, and I wouldn't want them
to come away with the impression that the method needed to be all
uppercase.