Refactor C bindings from nested function to module level variable
This is in support of adding PRIMA to SciPy. The nested function does
not work in musl on alpine. The reason is that alpine does not support
executable stacks. This internal function does work with gcc 14 since it
introduces -ftrampoline-impl=heap (as opposed to stack), but gcc 14 is
too cutting edge for scipy.

The module level variables are unfortunately not thread safe. Depending
on demand we could bring back the nested function under a flag, but I
think it's prudent to wait for demand for such a thing before
implementing it.