Update pybind11 to v3.0.1
This should fix the issue we were having with testing wheels on python
versions above 3.8. For those versions the default numpy was >=2.0, and
pybind11 did not support numpy 2.0 until v2.12. We were on 2.11, and
while upgrading to 2.12 would likely have been sufficient, I thought it
would be best if we could upgrade to the latest version so long as it
presented no major difficulties, and happily it did not.

Ref: #247