Copy pyprima code for fixed constraints to python bindings
Some minor modifications needed to be made. Notably pyprima's Result
object uses 'f' for the function value whereas the bindings use 'fun'.
Secondly, with the bindings we try to maintain the type that was
provided, and so we made to make a minor modification as we changed the
x0 that was provided to eliminate the fixed bounds.

Ref: #250